feat: move method and add account validation for wallet ids
